WAL files start at SequenceNumber::MIN (#7115)
## Usage and product changes

Sequenced records in the WAL are numbered starting from 1. Unsequenced
records use the number of the last sequenced record in the WAL.

A problem arises when an unsequenced record is written before any
sequenced records (e.g. server restart after WAL had been deleted). In
that case the index of the unsequenced record is forced to be zero.
However, the WAL shard file still believes it starts at 1, and when the
record header does not match that, the shard is believed to be
corrupted.

That is not an issue when sharding happens during an unsequenced write,
as the new shard filename uses the _record_ sequence number, not the
next expected sequence number. That does however mean that during a scan
through the WAL, if a shard file starts at the desired sequence number,
the previous shard must be scanned as well. (This is already the
implemented behaviour.)