Implement expression IR & evaluation (#7112)
## Usage and product changes
Implements translation from TypeQL, IR & evaluation of expressions.

## Implementation
#### Expression Compilation
The expression IR is a tree of expressions, where the leaves are either variables or constants and the internal nodes are operation & function symbols.
During compilation, this IR is resolved to `OpCode`s using the variable category & type information from type-inference. The output of compilation is a `CompiledExpression`, which can be executed by a stack based `evaluator` 
```
pub struct CompiledExpression {
    instructions: Vec<ExpressionOpCode>, // Instructions in the order to be executed.
    variables: Vec<Variable>,               // The list of variables in the order their corresponding load instructions appear
    constants: Vec<Value<'static>>,  // The list of constants in the order their corresponding load instructions appear 

    return_type: ExpressionValueType,        // The return type of the compiled expression - This is only needed to compile dependent instructions
}
```

#### Expression Evaluation
The effect of an `ExpressionOpCode` must be defined by an implementation of the `ExpressionInstruction` trait:
```
pub trait ExpressionInstruction: Sized {
    const OP_CODE: ExpressionOpCode;
    fn evaluate<'a>(state: &mut ExpressionEvaluationState<'a>) -> Result<(), ExpressionEvaluationError>;
}
```

where the argument is the state of our stack-based evaluator: 
```
pub struct ExpressionEvaluationState<'constants> {
    stack: Vec<ExpressionValue>,     // Holds the intermediate values for the stack-based evaluation
    variables: Box<[ExpressionValue]>, // Values of the input variables, in the reverse order of the CompiledExpression's variable_stack.
    next_variable_index: usize,   // The index of the next variable to be consumed.
    constants: &'constants [Value<'static>], // The constant_stack in the CompiledExpression
    next_constant_index: usize, // The index of the next constant to be consumed.
}
```

The evaluate function may read one or more items from the stack and push one or more items.

A number of macros & helper traits are provided to make implementing these easier (and less error-prone). These are all under the `builtins` module