Rearchitect packages on the read query path (#7121)
## Usage and product changes

We rearchitect the `//ir` and `//traversal` packages into several
different packges, to allow placing expressions and their compilation
and evaluation into a better cohesive achitecture.

We now have
```
//executor     expression execution and pattern execution
//compiler     which contains expression compilation, type inference, planning, and optimisation passes (future work)
//ir      which contains only the server-side representation and translation of the TypeQL into canonical form
```

We also have `//executor -> //compiler -> //ir`, so we each layer can
re-use data structures from the one below, and encompass logic for
building the more granular representation from the less granular one in
the more granular package. Note this is the inverted direction of
package dependencies we have normally, where we like the data flow to
match the package dependency direction!