Add data (thing) validation (#7136)
## Usage and product changes
We add data validation for schema migration (changes of the schema) and
data modifications.

Moreover, we add a number of additional operation time validations of
data for schema migration which only existed for transaction commits,
aiming to maximize the synchronization of schema and data of a database.
There is also a change in the cleanup behavior for relations and
attributes: now, nothing is hidden for queries until the commit time,
considering that we delete and hide data only on the commit time.

We refactor locking for things in `thing_manager`, making sure that not
only objects exist when we add edges to them, but also their interfaces
and other objects, connected to these objects through inheritance like
for `@unique` and `@card` constraints. The `@card` constraint, in its
turn, is now checked both in operation time for schema for all
operations with a potential effect on data's cardinality and in commit
time for data (in case something has been missed in the more complex
operation time checks).

Changes of previously existing checks:
1. Validations for types when an attribute type implicitly loses
@independent annotation or a relation type implicitly gains @cascade
annotation (which can lead to data deletion) are now throwing errors
only if there are existing instances for these types (including
subtypes).
2. We accidentally made the `create {root} type` idempotent, which is
incorrect. Now, if you try to create an existing entity type, it is an
error.

We also add the whole `concept` BDD tests package to the CI as it is
considered to be complete. From now on, we better aim not to break the
existing tests too much with new merges.