Introduce periodic fsync to guarantee commits are persisted to disk (#7149)
## Usage and product changes
Commits guarantee that changes have been persisted to the disk (using
`fsync`) before changes are visible and the commit is acknowledged. The
sync is done periodically and committing threads must wait for a signal
that the sync is complete. This avoids the performance penalty of each
commit doing its own sync, while providing guarantees that committed
data is not lost in the event of an OS crash.

## Implementation
* Introduce an `fsync` thread in the WAL.
* Introduce a `request_sync` API to the WAL, which returns a
`mpsc::receiver` which will receive a single `()` when the next fsync is
completed.