Fix greedy planner: variables are produced immediately (#7167)
## Release notes: product changes

Fixes a crash on a malformed intersection plan.

## Motivation

Previously the greedy planner could produce a double intersection. E.g.,
when the planner was given constraints `$x has $a`, `$x isa T`, `$a isa
A`, could then attempt to lump all the constraints into a single
intersection iterator producing `($x, $a)`.
The lowering of the plan to the executor, however, cannot support this
intersection (there is no corresponding executor, nor is one possible in
the current architecture), so it would e.g. intersect `$x has $a` and
`$x isa T` to produce `$x` and `$a`, then traverse `$a isa A`. As `$a`
was not registered as an input to the third constraint, it is unbound.
The executor then attempts to produce what is essentially a cartesian
product of `$a * $a`, encounters multiple values for `$a`, and crashes.
 
## Implementation

Mark variables as produced eagerly as the constraint is registered. This
does mean that the greedy planner is currently incapable of producing
intersections.