Disjunction support (#7205)
## Release notes: product changes

We introduce support for disjunctions in queries:

```php
match
    $person isa person;
    { $person has name $_; } or { $person has age $_; };
```

## Motivation

## Implementation

Proper support for disjunction including deduplication required overhaul
of the handling of selected variables in the planner and during
execution. In particular, rather than mapping variables to their
positions in the step output row 1-to-1, we introduce internal variables
that have no associated position, but can still be used for the purposes
of sorted joins or inline filters.

This PR also includes significant refactoring of the planner.