Implement tabling machinery for cyclic functions (#7224)
## Release notes: product changes
Introduces machinery needed to support finding the fixed-point of cyclic function calls. Cyclic functions now run, and return an incomplete set of results followed by an error. It is possible that the planner chooses a plan.

## Motivation
Reasonable waypoint towards full support.

## Implementation
Implements state for the tabled-functions as well as executors for tabled function calls.