Carry correct types into type list executor (#7243)
## Release notes: product changes

In cases where the list of types is dependent on the context (e.g. `{$t
label T;} or {$t label U;};`, the type list executor (role name,
explicit label, or kind) should only be producing the types in the list
defined by the constraint. That was handled correctly when the type
constraint was used as a post-check, but incorrectly during direct
execution.

## Motivation

## Implementation