Optimise has comparators and implement Executable Display (#7242)
## Release notes: product changes

We inline comparators into `HasReverse` executors, as a continuation of
#7233, which inlined comparators into `Isa` executors. Both
optimisations reduce the search space by using the value ranges provided
by constants in the query or previously executed query steps. We also
improve the query planner's statistics and cost function.

## Motivation

Traversals can do extra work scanning ranges of data when 1) we know a
range to search in eg, `$age > 18;` and 2) the data is sorted in an
order that we can take advantage of.

## Implementation

- Push value ranges into `HasReverse` execution, which causes the
iterators to skip data ranges that would never match the query
- Implement `Display` for `MatchExecutable`s, which allow much faster
debugging of slow or incorrect queries, which may always be caused by an
unusual query plan!
- Optimise the query planner statistics calculation, including making
in-memory costs non-zero.