Boxed large result Error variants (#7245)
## Release notes: product changes

We follow some best practices to box the large Error variants of Result
types, which should optimise both the stack size and the amount of
memory that has to be copied between function calls.

## Motivation

This refactor was prompted by spotting a non-trivial amount of time
doing `mem-copies` in flamegraphs. Although this is only 1 place we do
significant memory copies (others might be `clones` or collections that
are being reallocated), we expect that this optimisation has some impact
(and in fact, we measured maybe 1-2% improvement in TPCC).

## Implementation

- Where clippy warned us of a large `Err` variant of Results that are
being returned, we Box the Error type. This filters through to our error
`source` fields in error enums, function signatures, and construction of
errors. However, nothing is particularly interesting or fundamental.