Commit generated Cargo.tomls + Cargo.lock (#7240)
## Release notes: product changes

We commit the generated cargo manifests so that TypeDB can be built as a
normal cargo project without using Bazel.

Integration and unit tests can be run normally, and rust-analyzer
behaves correctly. Behaviour tests currently cannot be run without
Bazel, as the tests expect to find the feature files in
`bazel-typedb/external/typedb_behaviour`.

In addition, if during development the developer needs to depend on a
local clone of a dependency (e.g. making parallel changes to typeql),
the Cargo.toml would need to be temporarily manually adjusted to point
to the path dependency.

## Motivation

## Implementation

We also propagate the rename of vaticle bazel org prefix to typedb.