Implement structural equality (#7253)
## Release notes: product changes

We implement a form of Structural Equality, which check the equality of two queries, in Intermediate Representation format, which allows us to compare to queries for equality while ignoring user-written variable names and function names and constants. This is a building block for the next step, which will introduce a executable cache, allowing TypeDB to skip type inference and planning for queries that already exist in the cache.

## Motivation

We move towards the implementation of the query cache, which stores executable queries for a particular query structure. This will help amortise the cost of type inference and query planning for query pipelines. We start by implementing structural hashing and equality for all IR data structures.

## Implementation

We introduce a `StructuralEquality` trait, which is implemented by all IR data structures recursively, plus some base data types.