Query execution analyser (#7261)
## Release notes: product changes

We implement a useful debugging feature - a query analyzer. This is
similar to Postgres's `Explain Analyze`, which produces both the query
plan plus some details about the data that has flowed through the query
plan and the time at each step within it.

Example output:
```
Query profile[measurements_enabled=true]
  -----
  Stage or Pattern [id=0] - Match
    0. Sorted Iterator Intersection [bound_vars=[], output_size=1, sort_by=p0]
      [p0 isa ITEM] filter [] with (outputs=p0, )
    ==> batches: 158, rows: 10000, micros: 6407

    1. Sorted Iterator Intersection [bound_vars=[p0], output_size=2, sort_by=p1]
      Reverse[p1 rp p0 (role: __$2__)] filter [] with (inputs=p0, outputs=p1, checks=__$2__, )
    ==> batches: 854, rows: 39967, micros: 75716

  -----
  Stage or Pattern [id=1] - Reduce
    0. Reduction
    ==> batches: 1, rows: 10000, micros: 116035

  -----
  Stage or Pattern [id=2] - Insert
    0. Put attribute
    ==> batches: 10000, rows: 10000, micros: 5890

    1. Put has
    ==> batches: 10000, rows: 10000, micros: 54264
  ```

When disabled, profiling is a no-op (no strings are created, locks taken, or times measured), though there is still some cost associated with cloning Arcs containing the profiling data structures around. 

To enable query profiling, the easiest way (for now) is to enable TRACE logging level for the `executor` package, currrently configured in `//common/logger/logger.rs`:
```
.add_directive(LevelFilter::INFO.into())
// add:
// .add_directive("executor=trace".parse().unwrap())
```

Alternatively, just set the `enable` boolean to `true` in the `QueryProfile::new()` constructor.

## Motivation

To iterate more quickly debugging and finding performance issues, this kind of performance analyser, though it comes with some runtime overhead, is very useful to verify each stage of a query execution is producing the rows expected, and for verifying the amount of time taken to produce the rows. These together allow us to pinpoint 1) bad query plans, and 2) expensive operations.

## Implementation

We add a new struct `QueryProfile`, and its children, which are passed throughout the `//executor` package and used to record operations and timing against.

In the future, we should allow both returning this to the user in the Driver, and also enable or disable it per-query via the query options in the driver.