Fix greedy planner to sort on correct vars (#7263)
## Motivation

Currently, the greedy planner picks a plan in form of an ordering of
constraints and variables. An added constraint *produces* variables if
it relates those variables and they haven't been added to the plan at an
earlier stage.

The greedy planner picks constraints based on their "minimal" cost for a
directional lookup, that may sort on either of one or more variables
that the constraint relates. However, it does not record which variable
the lookup should be sorted on. This means, as it stands, produced
variables are added in random order after their constraint to the
ordered plan.

This PR aims to ensure that we sort on the variable for the appropriate
direction, by adding them first.

## Implementation

The previous implementation already had a notion of `sort_variable`
(renamed from `intersection_variable` since, in general, its really just
the sort variable for the step; moreover, the greedy plan essentially
never picks interesting intersections as it doesn't look ahead; and
produced variables are always cheaper to add.)

We simply set this `sort_variable` to be the variable corresponding to
the constraints direction.