Add more typing information into key prefixes to optimise storage lookups (#7262)
## Release notes: product changes

We incorporate more typing information into the storage lookup
operations, making them more longer and specific and reducing the range
of data that must be retrieved for any traversal operation. This should
make future RocksDB bloom filters more effective, and reduce the amount
of data that is read even when bloom filters are not available for the
prefix. Overall, we expect this to allow TypeDB to scale better,
particularly on larger-than-memory workloads.

## Motivation

We implement a feature that is already available in TypeDB 2.x: using
the destination type of a connection traversal (say, a 'has') to
restrict the range of data that must be searched. This should improve
performance and scalability.

## Implementation

We extend the prefixes used to do storage lookups to a consistent
length, where possible, of :
`[connection type][start concept][end concept type prefix]`
instead of
`[connection type][start concept]`. 

This should allow bloom filters to be more effective when they are
added, since we are searching for more specialised prefixes that will
hit fewer levels in the RocksDB LSM tree.

## Notes

- It is no long entirely clear that adding the `end concept type prefix`
is beneficial when the key range we need to scan is a range of prefixes,
rather than just within the one prefix: it seems that logical that Rocks
bloom filters only work exactly within 1 prefix, as they are used to
position iterators in the LSM tree files.
- If we want our iterators to support scanning beyond the initial prefix
without missing data, we will need to manually disable prefix filters:
```
Users can only use prefix bloom filter to read inside a specific prefix. If iterator is outside a prefix, the feature needs to be disabled for specific iterators to prevent wrong results:
```

from: https://github.com/facebook/rocksdb/wiki/Prefix-Seek

Thus, in RocksDB, we will need to be careful to make sure that the key
range we wish to scan in an iterator has a consistent prefix
(`key_range.start()[..bloom_prefix_length] ==
key_range.end()[..bloom_prefix_length]`)