Introduce 3.0 server diagnostics (#7288)
## Release notes: product changes
### Diagnostics
We introduce a diagnostics package to collect metrics related to server
usage. There are multiple potential clients of this package, a couple of
them being initially implemented:
* A web endpoint for pulling data (e.g. Prometheus) is bound at port
4104 by default and exposes diagnostics formatted for prometheus
(`http://localhost:4104/metrics?format=prometheus`) or as JSON
(`http://localhost:4104/metrics?format=JSON`) (done);
* A push http client for sending diagnostics data to the TypeDB
Diagnostics Service (done);
* PostHog (will be done in a near future).

If the reporting is turned off, we report a single batch of minimal
diagnostics information after 1 hour of runtime just to mark that a
server has been booted up. No user information is shared.
If the reporting is turned on, the reporting happens every hour
following [the rules set in
2.x](https://github.com/typedb/typedb/pull/7059).
If the development mode is active (details can be found below), no
reporting is executed, and the reporting config flag is ignored.

### Configuration
We add new CLI flags and for configuration:
1. `--diagnostics.monitoring.port 4014` for configuring the monitoring
server's port;
2. `--diagnostics.monitoring.enable false` for disabling and enabling
the monitoring server (true by default, thus called `enable` and
requires a boolean value);
3. `--diagnostics.reporting.metrics false` for disabling and enabling
the collected load and usage metrics' reporting (true by default,
ignored in the development mode).

An additional switch for the development mode has also been added in
order to use TypeDB's release binaries in CI.

### Development mode
We reintroduce the development mode. As in 2.x, this mode is a default
mode for local bazel (& cargo) builds and snapshots. It's turned off in
published releases but can be explicitly switched on through CLI, as
mentioned in the `Configuration` section.
**WARNING**: the development mode is used by the TypeDB's developers and
is not intended to be used in other programs. Be aware that its usage
can affect the server's performance and stability.