Propagate TypeQL syntax updates (#7289)
## Release notes: product changes

We mirror the changes from https://github.com/typedb/typeql/pull/383,
which is composed of 3 changes:

1) We now require that all assignments are preceded with a `let`
keyword:
```
match
 $p isa person, has monthly-salary $monthly;
 let $annual_salary = $monthly * 12;
```

2) We rename `long` to `integer`, to use a more familiar term for
integer values (`long` is very c-like!)
```
define
  attribute age, value integer;
```
This is still defined as a 64-bit signed integer in TypeDB's backend.

3) We make a consistent constructor syntax for instances:
```
insert
  $p isa person;      # create entity
  $f isa friendship (friend: $p);    # create relation with players
  $a isa age 20;      # create attribute with value
```

Previously, we would write this as:
```
insert 
  $p isa person;
  $f (friend: $p) isa friendship;
  $a isa age; $a == 20;
```

Note that the _anonymous relation constructor_ is now allowed like this:
```
insert
  friendship (friend: $p) ;
```
instead of:
```
insert
  (friend: $p) isa friendship;
```
 


## Motivation

For 1), the `let` binding makes it clearer to users when a attribute
value is a computed value within the query, and not being looked up from
storage.

For 2), we feel that the `integer` syntax will be clearer and more
familiar to users of most programming languages.

For 3), we resolve a long-standing inconsistency with giving
newly-created attributes a value. We already had the following syntactic
sugar for `has` clauses: `match $x isa person, has age 10;`. We make
this syntactic sugar as the basis of a consistent form for creating
entites, relations, and attributes.


## Implementation

We update all tests using the above syntaxes, and update the translation
phase to use the new TypeQL data structures.

---------

Co-authored-by: joshua <joshua@typedb.com>