Remove action metrics reporting to Posthog if reporting is disabled (#7301)
## Release notes: product changes
We remove action metrics reporting for all remote diagnostics endpoints
if reporting is disabled. Previously, while not sharing confidential
information, the server could still send the full action metrics
snapshot to one of the endpoints, which contained information about user
actions: numbers of opened connections, opened transactions, executed
queries, etc.

## Motivation
If reporting is disabled, we need to minimize the amount of data we
share, even if it's not personal. This approach reflects the traditional
diagnostics logic we had for the first Diagnostics Service endpoint.
## Implementation