Add Sentry for critical error reporting and fix panics on diagnostics initialization (#7308)
## Release notes: product changes
We add integration with Sentry to allow critical errors (e.g. `panic!`s)
reporting. This information will help us see and eliminate unexpected
TypeDB Server crashes.
Please use the `-diagnostics.reporting.errors` boolean option to disable
this feature if its work is undesirable (note that it will reduce the
efficiency of our maintenance).

Additionally, an overflow subtraction bug that sometimes affects
diagnostics initialization and leads to crashes is fixed.
 
## Motivation
The goal of the Sentry integration is to control unexpected errors,
which most likely indicate bugs or other issues introduced by
developers. It intentionally skips handled user errors as they are
reported to the diagnostics service as "user-related errors".

## Implementation
Based on [the official Sentry Rust setup
page](https://docs.sentry.io/platforms/rust/), the simplest way to
initialize Sentry correctly and not worry about threads and "sentry
hubs" is to leave Sentry in the `main` function instead of integrating
it inside of the `diagnostics` package (which was desired).
From the architecture point of view, it's not the cleanest solution, but
it is a decent approach from the perspective of how the Sentry crate
works in Rust. More comments are left in the PR diff.