Cartesian input fix (#7311)
## Release notes: product changes

We fix the issue where a join on an indexed relation would skip all
additional data within the join variable.

## Motivation

## Implementation

The `IndexedRelationExecutor` checks whether the input row contains a
value for the relation instance (regardless of mode). That value is then
used as a filter for the produced iterator. During a cartesian product,
instead of the input row, the previous instersection row (i.e., output
row) was passed into the executor during (re)opening. That meant that
the only accepted relation index tuples would have to come from the same
relation the first encountered one did.