Fix TypeDB reporting crashes when CA certificates are not found (#7318)
## Release notes: product changes
We fix TypeDB reporting crashes when CA certificates are not found on
the host. An additional Sentry (our crash reports endpoint) warning will
be reported, but it no longer affects the server's availability.

## Motivation

## Implementation
If reporting is not successful for either endpoints, a warning/error
(based on the cause) to Sentry will be reported, and the reporting for
this endpoint will stop. For simplicity, the job will be still activated
every 1 hour, but it will quit after a boolean flag check.

The next real attempt to report anything will happen only after a
restart.

The logic described above works for two endpoints separately. This way,
if only one endpoint fails (which should not be the case now), the other
one continues working as expected, reporting data every hour.