Cleanup todos (#7313)
## Release notes: product changes
Replaces usages of the todo macro with either errors which will be returned, or custom macros if the line is unreachable. This reduces server crashes due to panics when execution reaches the todo.

## Motivation
Avoid server crashes due to unimplemented code-paths.

## Implementation
Replaces usages of the todo macro (which would cause server crashes when hit) with either errors which will be returned, or custom macros if the line is unreachable. The custom macros (defined in `error.rs`) use an enum `UnimplementedFeature` which helps track unimplemented code by feature. This should be useful in the future when we're getting around to implementing them since we can just delete the variant and see what breaks.