Fix relates double specialization (#7330)
## Release notes: product changes
Fix the behavior of `relates` specialization, featuring:
* Unblocked double specialization:
```
define 
  relation family-relation relates member @abstract;
  relation parentship sub family relation, relates parent as member, relates child as member; # Good!
```  
* Fixed validations for multi-layered specializations:
```
define
  relation family-relation relates member @abstract, relates pet @abstract;
  relation parentship sub family relation, relates parent as member;
  relation fathership sub parentship, relates father as member; # Bad!
  relation fathership sub parentship, relates fathers-dog as pet; # Good!
```
* Better definition resolution and error messaging.
* Change the inner terminology for generated `relates` for
specialization: "root" and "non specializing" are substituted by
"explicit", and "specializing" is substituted by "implicit".

## Motivation
We fix https://github.com/typedb/typedb/issues/7322 and connected
issues.

## Implementation
Two steps:
**1.** 
The definable resolution was not correctly implemented for capabilities,
specifically for specialized relates. The new logic is the following:
* While we work with concept API, we **almost** always want to get all
capabilities, and the specialized capabilities should be returned with
the `@abstract` annotation. Thus, we do trust regular Concept API
methods everywhere.
* However, (re/un/)definitions are different. They need to find the
actually defined capabilities, not the ones being generated implicitly.
This way, a new set of methods is needed for these `non_specialised`
searches (decided to be explicit with the naming). Moreover, they need
to be straightforward, searching for a definition that a user would
think about.

With this simplified (the previous logic was quite complicated, at least
its reasoning, thus error-prone) search, the first part of the problem
has been solved. Now, define and redefine always check if they have a
**non-specialized** relates declared for them. Undefine uses
"transitive" search and searches for the closest non-specialized relates
when a role type is mentioned.

**2.**
The validations for specializations were not quite correct. They could
still allow some invalid usages of Concept API. Now, we check 2
conditions in multiple steps:
* The specialized `relates` should be had by my supertype (it was a
similar check, but generally for a role type, so it didn't quite work).
* The specialized `relates` should not be specializing (as before).

P.S. Sadly, the current state of Concept BDDs with a huge amount of
tests is not enough for multiple reasons (the first one is simply the
additional logic in definitions that can be buggy + a need to produce
more and more concept apis). I have some thoughts in my head which could
help us bring the old idea of auto-conversion to queries in reality
(probably only for schema queries: it will already be enough), but it
will take at least a weak to refactor and implement it. So we'll see...
In a year... Or more...